requirements.txt
Flask==3.0.2
python-dateutil==2.9.0.post0
qrcode==7.4.2
Pillow==10.4.0
